+++ LadybugPoker.xcodeproj/xcuserdata/rayoungnim.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -818,8 +818,8 @@
             filePath = "LadybugPoker/Views/GameRoom/DetailView/DetailTopView/PlayerBoardView.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "91"
-            endingLineNumber = "91"
+            startingLineNumber = "92"
+            endingLineNumber = "92"
             landmarkName = "playerBoard"
             landmarkType = "24">
          </BreakpointContent>
